requirements.txt
asgiref==3.7.2
Django==3.2.19
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
tweepy
textblob
mysql
python-dateutil